IT DOESN'T CRASH!!!! & some additions
