+++ Readme.md
@@ -3,7 +3,8 @@
 
 библиотека для (чуть более) удобной работы с KolibriOS, с похренитетом к оптимизации. Для всего и сразу.
 
-Сия библиотека представляет из себя прослойку абстракций, переименованных функций и прочей фигни, что упростит разработку приложений
+Сия библиотека представляет из себя прослойку абстракций, переименованных функций и прочей фигни, что упростит разработку приложений
+
 
 АХТУНГ! приложения использующие эту либу будут очень большими и медлеными.
 
@@ -32,6 +33,8 @@ mkdir build
 cmake ..
 make
 ```
+
+Под виндой всё работает отлично, для линукса же при линковке собранной библиотеки ошибка Archive index чото там
 
 
 ## Как использовать это
