+++ cmake/CMakeKOS32gcc.cmake
@@ -7,8 +7,8 @@
 message("Toolchain dir: " ${TOOLCHAIN_DIR})
 
 # Compiler
-SET(CMAKE_C_COMPILER	"${TOOLCHAIN_DIR}/bin/kos32-gcc${SUFFIX}")
-SET(CMAKE_CXX_COMPILER	"${TOOLCHAIN_DIR}/bin/kos32-g++${SUFFIX}")
+SET(CMAKE_C_COMPILER	"${TOOLCHAIN_DIR}/bin/kos32-gcc")
+SET(CMAKE_CXX_COMPILER	"${TOOLCHAIN_DIR}/bin/kos32-g++")
 SET(CMAKE_LINKER	"${TOOLCHAIN_DIR}/bin/kos32-ld${SUFFIX}")
 SET(CMAKE_AR	"${TOOLCHAIN_DIR}/bin/kos32-ar${SUFFIX}")
 set(CMAKE_C_COMPILER_AR	"${TOOLCHAIN_DIR}/bin/kos32-gcc-ar${SUFFIX}")
@@ -31,10 +31,15 @@ SET(SDK_DIR "${KOLIBRIOS_REPO}/contrib/sdk")
 set(LIBS_DIR ${SDK_DIR}/sources)
 
 #поключение newlib
-include_directories(${LIBS_DIR}/libstdc++-v3/include)
+include_directories(
+	${LIBS_DIR}/libstdc++-v3/include
+	${LIBS_DIR}/newlib/libc/include
+)
 
-#подключение libstdc++-v3
-include_directories(${LIBS_DIR}/newlib/libc/include)
-
+link_directories(
+	${SDK_DIR}/lib
+	${TOOLCHAIN_DIR}/mingw32/lib
+	${TOOLCHAIN_DIR}/lib
+)
 
 set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} -static -S -nostdlib -T ${SDK_DIR}/sources/newlib/app.lds -L ${TOOLCHAIN_DIR}/lib -L ${LIB_DIR}/lib -lgcc -lc.dll -lstdc++ -lsupc++")
